% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empb_beta_negbinomial.R
\name{empb_beta_negbinomial}
\alias{empb_beta_negbinomial}
\title{empb_beta_negbinomial}
\usage{
empb_beta_negbinomial(
  df,
  eta = 1,
  tol = 1e-08,
  maxIter = 10000,
  starting_rab = NULL,
  method = c("newton", "gdescent")
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
empb_beta_negbinomial
}
