% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empb_beta_negbinomial_c.R
\name{empb_beta_negbinomial_c}
\alias{empb_beta_negbinomial_c}
\title{empb_beta_negbinomial_c}
\usage{
empb_beta_negbinomial_c(
  df,
  eta = 1,
  tol = 1e-08,
  maxIter = 10000,
  starting_rab = NULL,
  method = c("newton", "gdescent")
)
}
\arguments{
\item{df}{data.frame object, containing at least columns 'x' containing non-negative integer values, and 'g' containing group labels.}

\item{eta}{positive numeric dampening parameter for Newton's method, gradient descent algorithm.}

\item{tol}{non-negative numeric tolerance parameter for exiting optimization algorithm.}

\item{maxIter}{positive integer setting maximum number of iterations for optimization algorithm.}

\item{starting_rab}{optional 3-long numeric vector, giving initial algorithm starting point for fitting empirical Bayes estimates for r, a, and b, respectively; default NULL.}

\item{method}{string controlling optimization method; default 'newton'.}
}
\value{
list object containing empirical Bayes (EMPB) estimates of r, a, and b hyperparameters, assuming df$x ~ nbinom(r, p_g), and p_g ~ beta(a, b), where 'p_g' denotes a group-level parameter.
}
\description{
empb_beta_negbinomial_c
}
\examples{
# Hello world
x = 2
}
