% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empb_gamma_poisson.R
\name{empb_gamma_poisson}
\alias{empb_gamma_poisson}
\title{empb_gamma_poisson}
\usage{
empb_gamma_poisson(
  df,
  eta = 1,
  tol = 1e-10,
  maxIter = 10000,
  starting_ab = NULL,
  method = c("newton", "gdescent")
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
empb_gamma_poisson
}
