% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape_binomial.R
\name{mape_binomial}
\alias{mape_binomial}
\title{mape_binomial}
\usage{
mape_binomial(df, a_prior, b_prior)
}
\arguments{
\item{df}{data.frame object, containing at least columns named 'x' containing non-negative integer values (number of successes), and 'n' containing non-negative integer values (number of trials).}

\item{a_prior}{positive numeric, giving prior parameter 'a', assuming binomial parameter p ~ beta(a, b).}

\item{b_prior}{positive numeric, giving prior parameter 'b', assuming binomial parameter p ~ beta(a, b).}
}
\value{
numeric, maximum a-posteriori estimate (MAPE) of binomial parameter p, assuming df$x ~ binom(p, df$n), and p ~ beta(a, b).
}
\description{
mape_binomial
}
\examples{
# Hello world
x = 2
}
