% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape_negbinomial.R
\name{mape_negbinomial}
\alias{mape_negbinomial}
\title{mape_negbinomial}
\usage{
mape_negbinomial(
  df,
  r_prior,
  p_prior,
  M_prior,
  eta = 1,
  tol = 1e-04,
  maxIter = 200
)
}
\arguments{
\item{maxIter}{}
}
\value{

}
\description{
mape_negbinomial
}
