% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape_poisson.R
\name{mape_poisson}
\alias{mape_poisson}
\title{mape_poisson}
\usage{
mape_poisson(df, a_prior, b_prior)
}
\arguments{
\item{df}{data.frame object, containing at least column named 'x' containing non-negative integer values.}

\item{a_prior}{positive numeric, giving prior parameter 'a', assuming Poisson parameter L ~ beta(a, b).}

\item{b_prior}{positive numeric, giving prior parameter 'b', assuming Poisson parameter L ~ beta(a, b).}
}
\value{
numeric, maximum a-posteriori estimate (MAPE) of Poisson parameter L, assuming df$x ~ poisson(L), and L ~ beta(a, b).
}
\description{
mape_poisson
}
\examples{
# Generate example data:
set.seed(31)
L = 5

# Number of experiments, i.e. rows in df:
numexps = 10

# Filling df with pseudo data; note the requisite column 'x':
df = data.frame('x' = rpois(numexps, L))

# Generating maximum a posteriori estimate (MAPE) solution for L:
L_fit = mape_poisson(df = df, a_prior = 1, b_prior = 1)

# Compare fitted values to known values:
cbind(L, L_fit)
}
