% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_binomial.R
\name{mle_binomial}
\alias{mle_binomial}
\title{mle_binomial}
\usage{
mle_binomial(df)
}
\arguments{
\item{df}{data.frame object, containing at least columns named 'x' containing non-negative integer values (number of successes), and 'n' containing non-negative integer values (number of trials)}
}
\value{
MLE of binomial distribution parameter p, assuming df$x ~ binom(p, df$n).
}
\description{
mle_binomial
}
\examples{
# Hello world
x = 2
}
