% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_negbinomial.R
\name{mle_negbinomial}
\alias{mle_negbinomial}
\title{mle_negbinomial}
\usage{
mle_negbinomial(
  df,
  eta = 0.001,
  lambda = 0.01,
  tol = 1e-04,
  maxIter = 10000,
  method = c("newton", "gdescent")
)
}
\arguments{
\item{method}{}
}
\value{

}
\description{
mle_negbinomial
}
