% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_poisson.R
\name{mle_poisson}
\alias{mle_poisson}
\title{mle_poisson}
\usage{
mle_poisson(df)
}
\arguments{
\item{df}{data.frame object, containing at least column named 'x' containing non-negative integer values.}
}
\value{
MLE of Poisson distribution parameter L, assuming df$x ~ poisson(L).
}
\description{
mle_poisson
}
\examples{
# Generate example data:
set.seed(31)
L = 5

# Number of experiments, i.e. rows in df:
numexps = 10

# Filling df with pseudo data; note the requisite column 'x':
df = data.frame('x' = rpois(numexps, L))

# Generating maximum likelihood estimate (MLE) solution for L:
L_fit = mle_poisson(df = df)

# Compare fitted values to known values:
cbind(L, L_fit)
}
