% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_gamma_poisson.R
\name{posterior_gamma_poisson}
\alias{posterior_gamma_poisson}
\title{posterior_gamma_poisson}
\usage{
posterior_gamma_poisson(
  df,
  ab_prior = NULL,
  dist = c("post", "pred"),
  Nsamp = 1000,
  ...
)
}
\arguments{
\item{df}{data.frame object, containing at least columns named 'x' containing non-negative integer values and 'g' containing group labels.}

\item{ab_prior}{length-2, positive numeric vector specifying prior hyperparameter values for a, b; if NULL, values fit by empirical Bayes (EMPB).}

\item{dist}{string specifying what type of samples to return: either 'post' (for samples from the posterior distribution), or 'pred' (for samples from the posterior-predictive distribution).}

\item{Nsamp}{positive integer, number of samples to generate per group.}

\item{...}{optional parameters to be passed to control EMPB convergence, in the case 'ab_prior' is NULL; see 'empb_gamma_poisson'.}
}
\value{
matrix of samples from the posterior (or posterior-predictive) distribution, where (named) columns are for group IDs included in df$g, and rows are samples; assuming df$x ~ binom(p_g, df$n), and p_g ~ beta(a, b), where 'p_g' denotes a group-level parameter.
}
\description{
posterior_gamma_poisson
}
\examples{
# Hello world
x = 2
}
