% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empb_beta_binomial.R
\name{empb_beta_binomial}
\alias{empb_beta_binomial}
\title{empb_beta_binomial}
\usage{
empb_beta_binomial(df, eta = 0.1, tol = 1e-05)
}
\arguments{
\item{tol}{}
}
\value{

}
\description{
empb_beta_binomial
}
