% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empb_gamma_poisson.R
\name{empb_gamma_poisson}
\alias{empb_gamma_poisson}
\title{empb_gamma_poisson}
\usage{
empb_gamma_poisson(
  df,
  eta = 1,
  tol = 1e-10,
  maxIter = 10000,
  starting_ab = NULL,
  method = c("newton", "gdescent")
)
}
\arguments{
\item{df}{data.frame object, containing at least columns 'x' containing non-negative integer values, and 'g' containing group labels.}

\item{eta}{positive numeric dampening parameter for Newton's method, gradient descent algorithm.}

\item{tol}{non-negative numeric tolerance parameter for exiting optimization algorithm.}

\item{maxIter}{positive integer setting maximum number of iterations for optimization algorithm.}

\item{starting_ab}{optional 2-long numeric vector, giving initial algorithm starting point for fitting empirical Bayes estimates for a and b; default NULL.}

\item{method}{string controlling optimization method; default 'newton'.}
}
\value{
list object containing empirical Bayes (EMPB) estimates of a, b hyperparameters, assuming df$x ~ poisson(L_g), and L_g ~ gamma(a, b), where 'L_g' denotes a group-level parameter.
}
\description{
empb_gamma_poisson
}
\examples{
# Hello world
x = 2
}
