% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empb_norm_negbinomial.R
\name{empb_norm_negbinomial}
\alias{empb_norm_negbinomial}
\title{empb_norm_negbinomial}
\usage{
empb_norm_negbinomial(
  df,
  lambda = 1,
  MLEeta = 0.001,
  EMPBeta = 0.001,
  tol = 0.1,
  maxIter = 10000,
  jitter = FALSE
)
}
\arguments{
\item{jitter}{}
}
\value{

}
\description{
empb_norm_negbinomial
}
