% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mape_binomial.R
\name{mape_binomial}
\alias{mape_binomial}
\title{mape_binomial}
\usage{
mape_binomial(df, a_prior, b_prior)
}
\arguments{
\item{df}{data.frame object, containing at least columns named 'x' containing non-negative integer values (number of successes), and 'n' containing non-negative integer values (number of trials).}

\item{a_prior}{positive numeric, giving prior parameter 'a', assuming binomial parameter p ~ beta(a, b).}

\item{b_prior}{positive numeric, giving prior parameter 'b', assuming binomial parameter p ~ beta(a, b).}
}
\value{
numeric, maximum a-posteriori estimate (MAPE) of binomial parameter p, assuming df$x ~ binom(p, df$n), and p ~ beta(a, b).
}
\description{
mape_binomial
}
\examples{
# Generate example data:
set.seed(31)
p = 0.3

# Number of experiments, i.e. rows in df:
numexps = 10

# Filling df with pseudo data; note the requisite columns 'n' and 'x':
n = 5 + rpois(numexps, 10)
x = rbinom(numexps, n, p)
df = data.frame('n' = n, 'x' = x)

# Generating maximum a posteriori estimate (MAPE) solution for p:
p_fit = mape_binomial(df = df, a_prior = 1, b_prior = 1)

# Compare fitted values to known values:
cbind(p, p_fit)
}
