% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_binomial.R
\name{mle_binomial}
\alias{mle_binomial}
\title{mle_binomial}
\usage{
mle_binomial(df)
}
\arguments{
\item{df}{data.frame object, containing at least columns named 'x' containing non-negative integer values (number of successes), and 'n' containing non-negative integer values (number of trials)}
}
\value{
MLE of binomial distribution parameter p, assuming df$x ~ binom(p, df$n).
}
\description{
mle_binomial
}
\examples{
# Generate example data:
set.seed(31)
p = 0.3

# Number of experiments, i.e. rows in df:
numexps = 10

# Filling df with pseudo data; note the requisite columns 'n' and 'x':
n = 5 + rpois(numexps, 10)
x = rbinom(numexps, n, p)
df = data.frame('n' = n, 'x' = x)

# Generating maximum likelihood estimate (MLE) solution for p:
p_fit = mle_binomial(df = df)

# Compare fitted values to known values:
cbind(p, p_fit)
}
