% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mle_poisson.R
\name{mle_poisson}
\alias{mle_poisson}
\title{mle_poisson}
\usage{
mle_poisson(df)
}
\arguments{
\item{df}{data.frame object, containing at least column named 'x' containing non-negative integer values.}
}
\value{
MLE of Poisson distribution parameter L, assuming df$x ~ poisson(L).
}
\description{
mle_poisson
}
\examples{
# Hello world
x = 2
}
