% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posterior_beta_binomial.R
\name{posterior_beta_binomial}
\alias{posterior_beta_binomial}
\title{posterior_beta_binomial}
\usage{
posterior_beta_binomial(
  df,
  ab_prior = NULL,
  dist = c("post", "pred"),
  pred_n = NULL,
  Nsamp = 1000,
  ...
)
}
\arguments{
\item{df}{data.frame object, containing at least columns named 'x' containing non-negative integer values (number of successes), 'n' containing non-negative integer values (number of trials), and 'g' containing group labels.}

\item{ab_prior}{length-2, positive numeric vector specifying prior hyperparameter values for a, b; if NULL, values fit by empirical Bayes.}

\item{dist}{string specifying what type of samples to return: either 'post' (for samples from the posterior distribution), or 'pred' (for samples from the posterior-predictive distribution).}

\item{pred_n}{positive integer, of length 1 or length equal to number of unique elements in df$g: 'size' of binomial samples if generating samples from posterior-predictive distribution.}

\item{Nsamp}{positive integer, number of samples to generate per group.}

\item{...}{optional parameters to be passed to control EMPB convergence, in the case 'ab_prior' is NULL; see 'empb_beta_binomial'.}
}
\value{
list object containing empirical Bayes (EMPB) estimates of a, b hyperparameters, assuming df$x ~ binom(p_g, df$n), and p_g ~ beta(a, b), where 'p_g' denotes a group-level parameter.
}
\description{
posterior_beta_binomial
}
\examples{
# Hello world
x = 2
}
